Add files via upload
A bus management system helps to register and keep information about all bus. A summary can also be generated from this system. Reports can be generated from the database records so that the bus can be tracked and the budget can be seen. It provides facility to instant and quick reservation.

The main objective of this entire activity is to automate the activities like :
Booking
Vehicle details
Passenger details
Bill details

Following Modules are Used in this project :
Tkinter